+++ javascript/packages/eslint-config-chotot/next.js
@@ -1,0 +1,12 @@
+/**
+ * In .eslintrc.js, declare like this:
+ * {
+ *   extends: [ 'chotot', 'chotot/next' ]
+ * }
+ */
+module.exports = {
+  extends: [
+    './rules/next.js',
+  ].map(require.resolve),
+  rules: {}
+};