+++ javascript/packages/eslint-config-chotot/rules/next.js
@@ -1,0 +1,20 @@
+/**
+ * Set of rules to override / sepcific for Next.js framework
+ */
+module.exports = {
+  plugins: [
+    'react',
+  ],
+
+  parserOptions: {
+    ecmaFeatures: {
+      jsx: true,
+    },
+  },
+
+  rules: {
+    // This is not required in Next.js project
+    // https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/react-in-jsx-scope.md
+    'react/react-in-jsx-scope': 'off',
+  },
+};