+++ javascript/packages/eslint-config-chotot/CHANGELOG.md
@@ -1,4 +1,9 @@
+18.1.0 / 2020-02-11
+===================
+- Add `chotot/next` configs specialized for projects using Next.js framework
+- Add `babel-eslint` as peer dependency so that it show up in `npm info "eslint-config-chotot@latest" peerDependencies`
+
 18.0.0 / 2019-12-19
-==================
+===================
 - Initial version, forked from `eslint-config-airbnb`@18.0.1
 - Version continue from the forked package for reference